+++ contrib/scripts/release.sh
@@ -7,6 +7,7 @@ args=(
     "-p nostr-database"
     "-p nostr-lmdb"
     "-p nostr-mls-storage"
+    "-p nostr-mls-memory-storage"
     "-p nostr-ndb"
     "-p nostr-indexeddb"
     "-p nostr-keyring"